# docs/backend/openapi.yaml
openapi: 3.1.0
info:
  title: Mosaic API
  version: 0.7.0-phase7
  description: |
    Hosted workspace, provider foundation, publishing, Asset, browser-session, and SDK-delivery API.
    RevenueCat server-connected integrations use least-privilege v2 secret keys. Secrets
    are accepted only on create, rotate, and reconnect and are never returned.
servers:
  - url: /
security:
  - BrowserSession: []
paths:
...
  /v1/sdk/events/batch:
    post:
      security: [{ PublicSDKKey: [] }]
      operationId: ingestAnalyticsEventBatch
      tags: [Analytics]
      description: The enclosing Analytics Event contract version must exactly match every eventSchemaVersion in the batch; mixed-version batches reject atomically.
      requestBody: { required: true, content: { application/json: { schema: { $ref: "#/components/schemas/AnalyticsEventBatch" } } } }
      responses:
        "200": { description: Per-event ingestion outcomes., content: { application/json: { schema: { $ref: "#/components/schemas/AnalyticsIngestionResult" } } } }
        "401": { $ref: "#/components/responses/Error" }
        "409": { $ref: "#/components/responses/Error" }
        "422": { $ref: "#/components/responses/Error" }
        "429": { $ref: "#/components/responses/Error" }
  /v1/projects/{projectId}/environments/{environmentId}/analytics/settings:
    parameters: [{ $ref: "#/components/parameters/ProjectID" }, { $ref: "#/components/parameters/EnvironmentID" }]
    get:
      operationId: getAnalyticsSettings
      tags: [Analytics]
      responses: { "200": { description: Analytics settings., content: { application/json: { schema: { $ref: "#/components/schemas/AnalyticsSettingsEnvelope" } } } }, "403": { $ref: "#/components/responses/Error" }, "404": { $ref: "#/components/responses/Error" } }
    put:
      operationId: updateAnalyticsSettings
      tags: [Analytics]
      requestBody: { required: true, content: { application/json: { schema: { $ref: "#/components/schemas/UpdateAnalyticsSettingsRequest" } } } }
      responses: { "200": { description: Updated analytics settings., content: { application/json: { schema: { $ref: "#/components/schemas/AnalyticsSettingsEnvelope" } } } }, "403": { $ref: "#/components/responses/Error" }, "422": { $ref: "#/components/responses/Error" } }
  /v1/projects/{projectId}/environments/{environmentId}/analytics/overview:
    get:
      operationId: getAnalyticsOverview
      tags: [Analytics]
      parameters: [{ $ref: "#/components/parameters/ProjectID" }, { $ref: "#/components/parameters/EnvironmentID" }, { $ref: "#/components/parameters/AnalyticsFrom" }, { $ref: "#/components/parameters/AnalyticsTo" }, { $ref: "#/components/parameters/AnalyticsTimezone" }, { $ref: "#/components/parameters/AnalyticsMetricBasis" }, { $ref: "#/components/parameters/AnalyticsPlatform" }, { $ref: "#/components/parameters/AnalyticsLocale" }, { $ref: "#/components/parameters/AnalyticsApplicationVersion" }]
      responses: { "200": { $ref: "#/components/responses/AnalyticsResult" }, "403": { $ref: "#/components/responses/Error" }, "422": { $ref: "#/components/responses/Error" } }
  /v1/projects/{projectId}/environments/{environmentId}/analytics/funnels/{funnel}:
    get:
      operationId: getAnalyticsFunnel
      tags: [Analytics]
      parameters: [{ $ref: "#/components/parameters/ProjectID" }, { $ref: "#/components/parameters/EnvironmentID" }, { name: funnel, in: path, required: true, schema: { type: string, enum: [placements, paywalls, products, purchases] } }, { $ref: "#/components/parameters/AnalyticsFrom" }, { $ref: "#/components/parameters/AnalyticsTo" }, { $ref: "#/components/parameters/AnalyticsTimezone" }, { $ref: "#/components/parameters/AnalyticsMetricBasis" }, { $ref: "#/components/parameters/AnalyticsPlatform" }, { $ref: "#/components/parameters/AnalyticsLocale" }, { $ref: "#/components/parameters/AnalyticsApplicationVersion" }]
      responses: { "200": { $ref: "#/components/responses/AnalyticsResult" }, "404": { $ref: "#/components/responses/Error" }, "422": { $ref: "#/components/responses/Error" } }
  /v1/projects/{projectId}/environments/{environmentId}/analytics/paywall-version-comparison:
    get: { operationId: compareAnalyticsPaywallVersions, tags: [Analytics], parameters: [{ $ref: "#/components/parameters/ProjectID" }, { $ref: "#/components/parameters/EnvironmentID" }, { $ref: "#/components/parameters/AnalyticsFrom" }, { $ref: "#/components/parameters/AnalyticsTo" }, { $ref: "#/components/parameters/AnalyticsTimezone" }, { $ref: "#/components/parameters/AnalyticsMetricBasis" }, { $ref: "#/components/parameters/AnalyticsPlatform" }, { $ref: "#/components/parameters/AnalyticsLocale" }, { $ref: "#/components/parameters/AnalyticsApplicationVersion" }], responses: { "200": { $ref: "#/components/responses/AnalyticsResult" }, "422": { $ref: "#/components/responses/Error" } } }
  /v1/projects/{projectId}/environments/{environmentId}/analytics/provider-errors:
    get: { operationId: getAnalyticsProviderErrors, tags: [Analytics], parameters: [{ $ref: "#/components/parameters/ProjectID" }, { $ref: "#/components/parameters/EnvironmentID" }, { $ref: "#/components/parameters/AnalyticsFrom" }, { $ref: "#/components/parameters/AnalyticsTo" }, { $ref: "#/components/parameters/AnalyticsTimezone" }, { $ref: "#/components/parameters/AnalyticsMetricBasis" }, { $ref: "#/components/parameters/AnalyticsPlatform" }, { $ref: "#/components/parameters/AnalyticsLocale" }, { $ref: "#/components/parameters/AnalyticsApplicationVersion" }], responses: { "200": { $ref: "#/components/responses/AnalyticsResult" }, "422": { $ref: "#/components/responses/Error" } } }
  /v1/projects/{projectId}/environments/{environmentId}/analytics/product-availability-failures:
    get: { operationId: getAnalyticsProductAvailabilityFailures, tags: [Analytics], parameters: [{ $ref: "#/components/parameters/ProjectID" }, { $ref: "#/components/parameters/EnvironmentID" }, { $ref: "#/components/parameters/AnalyticsFrom" }, { $ref: "#/components/parameters/AnalyticsTo" }, { $ref: "#/components/parameters/AnalyticsTimezone" }, { $ref: "#/components/parameters/AnalyticsMetricBasis" }, { $ref: "#/components/parameters/AnalyticsPlatform" }, { $ref: "#/components/parameters/AnalyticsLocale" }, { $ref: "#/components/parameters/AnalyticsApplicationVersion" }], responses: { "200": { $ref: "#/components/responses/AnalyticsResult" }, "422": { $ref: "#/components/responses/Error" } } }
  /v1/projects/{projectId}/environments/{environmentId}/analytics/breakdowns/{dimension}:
    get:
      operationId: getAnalyticsBreakdown
      tags: [Analytics]
      parameters: [{ $ref: "#/components/parameters/ProjectID" }, { $ref: "#/components/parameters/EnvironmentID" }, { name: dimension, in: path, required: true, schema: { type: string, enum: [platforms, locales] } }, { $ref: "#/components/parameters/AnalyticsFrom" }, { $ref: "#/components/parameters/AnalyticsTo" }, { $ref: "#/components/parameters/AnalyticsTimezone" }, { $ref: "#/components/parameters/AnalyticsMetricBasis" }, { $ref: "#/components/parameters/AnalyticsPlatform" }, { $ref: "#/components/parameters/AnalyticsLocale" }, { $ref: "#/components/parameters/AnalyticsApplicationVersion" }]
      responses: { "200": { $ref: "#/components/responses/AnalyticsResult" }, "404": { $ref: "#/components/responses/Error" }, "422": { $ref: "#/components/responses/Error" } }
  /v1/projects/{projectId}/environments/{environmentId}/analytics/freshness:
    get: { operationId: getAnalyticsFreshness, tags: [Analytics], parameters: [{ $ref: "#/components/parameters/ProjectID" }, { $ref: "#/components/parameters/EnvironmentID" }, { $ref: "#/components/parameters/AnalyticsFrom" }, { $ref: "#/components/parameters/AnalyticsTo" }, { $ref: "#/components/parameters/AnalyticsTimezone" }, { $ref: "#/components/parameters/AnalyticsMetricBasis" }], responses: { "200": { $ref: "#/components/responses/AnalyticsResult" }, "422": { $ref: "#/components/responses/Error" } } }
  /v1/projects/{projectId}/environments/{environmentId}/analytics/exports:
    post:
      operationId: createAnalyticsEventExport
      tags: [Analytics]
      parameters: [{ $ref: "#/components/parameters/ProjectID" }, { $ref: "#/components/parameters/EnvironmentID" }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: "#/components/schemas/CreateAnalyticsEventExportRequest" } } } }
      responses: { "202": { $ref: "#/components/responses/AnalyticsJob" }, "403": { $ref: "#/components/responses/Error" }, "422": { $ref: "#/components/responses/Error" } }
...
components:
  securitySchemes:
    BrowserSession:
      type: apiKey
      in: cookie
      name: mosaic_session
      description: Opaque HttpOnly SameSite=Lax browser session. Cross-origin dashboard requests use credentials mode include.
    PublicSDKKey:
      type: http
      scheme: bearer
      bearerFormat: Mosaic public SDK key
  parameters:
...
    ProjectID: { in: path, name: projectId, required: true, schema: { type: string } }
    EnvironmentID: { in: path, name: environmentId, required: true, schema: { type: string } }
...
    AnalyticsFrom: { in: query, name: from, required: true, schema: { $ref: "#/components/schemas/Timestamp" } }
    AnalyticsTo: { in: query, name: to, required: true, schema: { $ref: "#/components/schemas/Timestamp" } }
    AnalyticsTimezone: { in: query, name: timezone, required: true, schema: { type: string, example: UTC } }
    AnalyticsMetricBasis: { in: query, name: metricBasis, required: true, schema: { type: string, const: event_count } }
    AnalyticsPlatform: { in: query, name: platform, schema: { type: string, enum: [ios, android] } }
    AnalyticsLocale: { in: query, name: locale, schema: { type: string } }
    AnalyticsApplicationVersion: { in: query, name: applicationVersion, schema: { type: string } }
  responses:
    Error:
      description: Stable machine-readable failure.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
    AnalyticsResult: { description: Event-count analytics result., content: { application/json: { schema: { $ref: "#/components/schemas/AnalyticsResultEnvelope" } } } }
    AnalyticsJob: { description: Asynchronous analytics job., content: { application/json: { schema: { $ref: "#/components/schemas/AnalyticsJobEnvelope" } } } }
...
  schemas:
...
    AnalyticsEventBatch:
      type: object
      additionalProperties: false
      required: [analyticsEventContractVersion, batchId, sentAt, events]
      properties:
        analyticsEventContractVersion: { type: string, enum: ["1", "2"], description: Must exactly equal every enclosed eventSchemaVersion. }
        batchId: { type: string }
        sentAt: { $ref: "#/components/schemas/Timestamp" }
        events: { type: array, minItems: 1, maxItems: 100, items: { type: object, additionalProperties: true, description: Canonical Analytics Event document matching analyticsEventContractVersion exactly. } }
    AnalyticsEventResult:
      type: object
      additionalProperties: false
      required: [eventId, status]
      properties:
        eventId: { type: string }
        status: { type: string, enum: [accepted, duplicate, permanently_rejected, retryable] }
        code: { type: string }
    AnalyticsIngestionResult:
      type: object
      additionalProperties: false
      required: [analyticsEventContractVersion, batchId, receivedAt, results]
      properties:
        analyticsEventContractVersion: { type: string, enum: ["1", "2"] }
        batchId: { type: string }
        receivedAt: { $ref: "#/components/schemas/Timestamp" }
        results: { type: array, items: { $ref: "#/components/schemas/AnalyticsEventResult" } }
    AnalyticsSettings:
      type: object
      additionalProperties: false
      required: [projectId, environmentId, collectionEnabled, rawRetentionDays, updatedAt]
      properties:
        projectId: { type: string }
        environmentId: { type: string }
        collectionEnabled: { type: boolean, default: false }
        rawRetentionDays: { type: integer, minimum: 30, maximum: 730, default: 180 }
        updatedAt: { $ref: "#/components/schemas/Timestamp" }
    UpdateAnalyticsSettingsRequest:
      type: object
      additionalProperties: false
      required: [collectionEnabled, rawRetentionDays]
      properties:
        collectionEnabled: { type: boolean }
        rawRetentionDays: { type: integer, minimum: 30, maximum: 730 }
    AnalyticsFreshness:
      type: object
      additionalProperties: false
      required: [lateEventPolicy]
      properties:
        latestReceivedAt: { $ref: "#/components/schemas/Timestamp" }
        latestAggregatedAt: { $ref: "#/components/schemas/Timestamp" }
        lateEventPolicy: { type: string }
    AnalyticsMetric:
      type: object
      additionalProperties: false
      required: [id, numerator, basis, authority, attributionWindow, timezone, definition]
      properties:
        id: { type: string }
        value: { type: number }
        numerator: { type: integer, format: int64 }
        denominator: { type: integer, format: int64 }
        basis: { type: string, const: event_count }
        authority: { type: string, enum: [client_observed, trusted_server, provider_confirmed] }
        attributionWindow: { type: string, const: 24h }
        timezone: { type: string }
        definition: { type: string }
        warnings: { type: array, items: { type: string } }
        dimensions: { type: object, additionalProperties: { type: string } }
    AnalyticsResult:
      type: object
      additionalProperties: false
      required: [metrics, freshness]
      properties:
        metrics: { type: array, items: { $ref: "#/components/schemas/AnalyticsMetric" } }
        freshness: { $ref: "#/components/schemas/AnalyticsFreshness" }
    CreateAnalyticsEventExportRequest:
      type: object
      additionalProperties: false
      required: [from, to, format]
      properties:
        from: { $ref: "#/components/schemas/Timestamp" }
        to: { $ref: "#/components/schemas/Timestamp" }
        format: { type: string, enum: [ndjson, csv] }
...
    AnalyticsJob:
      type: object
      additionalProperties: false
      required: [id, kind, status, createdAt, updatedAt]
      properties:
        id: { type: string }
        kind: { type: string, enum: [events, application_user, installation] }
        status: { type: string, enum: [queued, leased, recomputing, completed, failed, expired] }
        format: { type: string, enum: [ndjson, csv] }
        rowCount: { type: integer, format: int64 }
        byteLength: { type: integer, format: int64 }
        affectedEventCount: { type: integer, format: int64 }
        affectedSessionCount: { type: integer, format: int64 }
        expiresAt: { $ref: "#/components/schemas/Timestamp" }
        createdAt: { $ref: "#/components/schemas/Timestamp" }
        updatedAt: { $ref: "#/components/schemas/Timestamp" }
    AnalyticsSettingsEnvelope: { type: object, required: [data], properties: { data: { $ref: "#/components/schemas/AnalyticsSettings" } } }
    AnalyticsResultEnvelope: { type: object, required: [data], properties: { data: { $ref: "#/components/schemas/AnalyticsResult" } } }
...
    AnalyticsJobEnvelope: { type: object, required: [data], properties: { data: { $ref: "#/components/schemas/AnalyticsJob" } } }
...
    Timestamp: { type: string, format: date-time }
...
    ErrorEnvelope:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error:
          type: object
          additionalProperties: false
          required: [code, message]
          properties:
            code: { type: string }
            message: { type: string }
            fields: { type: object, additionalProperties: { type: array, items: { type: string } } }
            details: { type: object, additionalProperties: true }
            requestId: { type: string }
...